[move] : LongTextField 컴포넌트 index.tsx 생성 후 폴더화
