+++ packages/shared/src/components/longTextField/index.tsx
@@ -1,0 +1,1 @@
+export { default as LongTextField } from './LongTextField';
